# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "B2B Partner OAuth 2.0 Movie API Specification",
    "description": "B2B 파트너 연동을 위한 OAuth 2.0 Client Credentials 기반 영화 REST API 규격서입니다. 외부 개발사에 이 스펙(OpenAPI JSON/YAML)을 그대로 전달하여 자동 코드 생성(Swagger Codegen) 및 클라이언트 연동에 활용할 수 있습니다.",
    "version": "1.0.0",
    "contact": {
      "name": "B2B API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "OAuth 2.0 Authentication",
      "description": "B2B 업체 Access Token 발급 및 인증"
    },
    {
      "name": "B2B Movie Resources",
      "description": "파트너 업체 영화 데이터 조회 및 등록"
    },
    {
      "name": "Company Management",
      "description": "업체 프로필 및 실시간 API 호출 트래킹 로그"
    }
  ],
  "paths": {
    "/api/oauth/token": {
      "post": {
        "tags": ["OAuth 2.0 Authentication"],
        "summary": "B2B 업체 Access Token 발급 (Client Credentials Grant)",
        "description": "발급받은 업체 Client ID와 Client Secret으로 24시간 유효한 업체 전용 Bearer Access Token을 발급받습니다.",
        "operationId": "issueAccessToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              },
              "example": {
                "grant_type": "client_credentials",
                "client_id": "partner_minstudio",
                "client_secret": "secret_minstudio_key123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access Token 발급 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                },
                "example": {
                  "access_token": "comp_at_89f21a...",
                  "token_type": "Bearer",
                  "expires_in": 86400,
                  "scope": "read,write",
                  "company_name": "민스튜디오 엔터테인먼트",
                  "company_id": 1
                }
              }
            }
          },
          "401": {
            "description": "Client ID 또는 Client Secret 인증 실패"
          }
        }
      }
    },
    "/api/v1/movies": {
      "get": {
        "tags": ["B2B Movie Resources"],
        "summary": "업체 영화 목록 조회",
        "description": "영화 목록을 조회합니다. Header에 Bearer 업체 토큰을 전달하면 해당 업체가 배급하는 영화 데이터만 필터링하여 제공합니다.",
        "operationId": "getMovies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "영화 목록 조회 성공",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "[민스튜디오 엔터테인먼트] 업체 전용 영화 목록 조회 완료",
                  "company": {
                    "id": 1,
                    "name": "민스튜디오 엔터테인먼트",
                    "clientId": "partner_minstudio"
                  },
                  "count": 2,
                  "data": [
                    {
                      "MOVIE_ID": 1,
                      "COMPANY_ID": 1,
                      "TITLE": "인터스텔라 (민스튜디오 배급)",
                      "ORIGINAL_TITLE": "Interstellar",
                      "RUNNING_TIME": 169,
                      "PLOT": "시공간을 탐험하는 인류의 이야기",
                      "COMPANY_NAME": "민스튜디오 엔터테인먼트"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않거나 만료된 Access Token"
          }
        }
      },
      "post": {
        "tags": ["B2B Movie Resources"],
        "summary": "자사 신규 영화 데이터 등록",
        "description": "Bearer 업체 토큰으로 인증하여 파트너 소유의 신규 영화를 등록합니다.",
        "operationId": "createMovie",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieCreateRequest"
              },
              "example": {
                "title": "오펜하이머 (민스튜디오 배급)",
                "original_title": "Oppenheimer",
                "running_time": 180,
                "plot": "세상을 바꾼 천재 과학자의 이야기"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "영화 등록 성공",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "[민스튜디오 엔터테인먼트] 소유의 신규 영화 데이터가 성공적으로 등록되었습니다.",
                  "movie": {
                    "company_name": "민스튜디오 엔터테인먼트",
                    "title": "오펜하이머 (민스튜디오 배급)",
                    "original_title": "Oppenheimer",
                    "running_time": 180
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization 헤더 누락 또는 토큰 인증 실패"
          }
        }
      }
    },
    "/api/v1/companies/me": {
      "get": {
        "tags": ["Company Management"],
        "summary": "내 업체 프로필 & API 호출 실시간 로그 조회",
        "description": "현재 토큰의 업체 계정 정보, 일일 API 호출 한도 및 최근 API 호출 실시간 트래킹 로그를 조회합니다.",
        "operationId": "getCompanyProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 및 호출 로그 조회 성공",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "COMPANY_ID": 1,
                    "COMPANY_NAME": "민스튜디오 엔터테인먼트",
                    "BUSINESS_NUMBER": "123-45-67890",
                    "CLIENT_ID": "partner_minstudio",
                    "STATUS": "ACTIVE",
                    "DAILY_LIMIT": 10000
                  },
                  "call_logs": [
                    {
                      "ENDPOINT": "/api/v1/movies",
                      "HTTP_METHOD": "GET",
                      "STATUS_CODE": 200,
                      "CREATED_AT": "2026-08-06T11:30:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT/OAuth2 Token",
        "description": "POST /api/oauth/token 에서 발급받은 Bearer Access Token을 입력합니다. (예: Bearer comp_at_...)"
      }
    },
    "schemas": {
      "TokenRequest": {
        "type": "object",
        "required": ["grant_type", "client_id", "client_secret"],
        "properties": {
          "grant_type": {
            "type": "string",
            "enum": ["client_credentials"],
            "description": "OAuth 2.0 Grant 방식 (client_credentials 고정)"
          },
          "client_id": {
            "type": "string",
            "description": "파트너 업체 Client ID"
          },
          "client_secret": {
            "type": "string",
            "description": "파트너 업체 Client Secret"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "24시간 유효한 Bearer Access Token"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "example": 86400
          },
          "scope": {
            "type": "string",
            "example": "read,write"
          },
          "company_name": {
            "type": "string",
            "example": "민스튜디오 엔터테인먼트"
          }
        }
      },
      "MovieCreateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "description": "영화 제목"
          },
          "original_title": {
            "type": "string",
            "description": "영화 원제"
          },
          "running_time": {
            "type": "integer",
            "description": "상영 시간 (분 단위)"
          },
          "plot": {
            "type": "string",
            "description": "영화 줄거리 상세"
          }
        }
      }
    }
  }
}